test(desktop): split agent tree specs
